Fixed #1709 issue on map inteq_colossus.dmm (#1796)
## About The Pull Request
I swapped gas tanks on map inteq_colossus. There was a problem with
them, that instead of 75% nitrogen and 25% oxygen
solution, it was 75% oxygen and 25% nitrogen. That explains issue #1709 

## Why It's Good For The Game

It's fixes an issue that can ruin a game.

## Changelog

:cl: Davchik
fix: swapped two gas tanks (oxygen and nitrogen) on map
inteq_colssus.dmm
/:cl: