Fixes Wikibooks (#1829)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Port of: yogstation13/Yogstation#11662

Gets wikibooks working again, hopefully. This was a pain.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->


![image](https://user-images.githubusercontent.com/29362068/224886450-0f297943-c6c2-4891-914f-e8c39f86a1ee.png)

## Why It's Good For The Game
Wikibooks are a good way of showing wiki info IC. Or so I'm told.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Wikibooks should now work.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: alexkar598 <25136265+alexkar598@users.noreply.github.com>