Adds new lizard emotes (#1877)
## About The Pull Request
Adds *hiss and *weh for lizards

Hiss is from https://freesound.org/people/crcavol/sounds/154636/,
slightly edited
Weh is from skyrat
## Why It's Good For The Game
1. Lizardy emote
2. weh
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
soundadd: Adds *hiss and *weh for Sarathi.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->