fixes jungle_medtech_outbreak not spawning (#1910)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
at some point the DATUM for the ruin was removed, causing it to not
spawn at all. This PR readds the datum!
it also changes the outside turfs to the "lit" variant, and switches the
lattices to the fake version
and it adds some wooden lining 

![2023 04 17-11 48
05](https://user-images.githubusercontent.com/79304582/232449449-3c765dea-456f-466d-b943-cd0a655e1ca3.png)


![image](https://user-images.githubusercontent.com/79304582/232450065-31a909e8-00fa-4457-a885-c734652fb60b.png)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
good ruins should, infact, spawn!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: fixed jungle_medtech_outbreak
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->