Reroutes the Skipper's TEG cold-loop pipes (#1963)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Routes the Skipper's cool-loop pipes to go AROUND the captain's office
and follow through open-turf plating. Also moves the potted plant and
freezer to accommodate (why is the engine freezer in the cryogenics
area?). Also also removes the floor from underneath the latices in
engineering.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Good for engineers, as we won't have to break down **reinforced walls**
to make modifications to pipes.

### Before


![image](https://github.com/shiptest-ss13/Shiptest/assets/30960302/f15b13bf-c214-4c60-969a-2ba801c08090)

### After


![image](https://github.com/shiptest-ss13/Shiptest/assets/30960302/56466958-b895-480c-993e-16792c5db337)


<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: moved the skipper's pipes
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->