Removes crew manifest from PDA, Modcomp, AI, PAI (#1987)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I do not think this information should be available to living players.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Having an exact print out of exactly who is on someone's ship is very
bad for any conflict scenario and feels a little off. We aren't on a
station, every ship doesn't need a print out of other ships. You can
call ships on a holopad for a similar effect without being so verbose.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: PDAs, Modcomps, AIs, and pAIs no longer have a crew manifest
option.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->