Flyswatter sprite changes (#2000)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Redone flyswatter sprites, including inhands. Mostly done as a personal
test run of how to contribute.

![image](https://github.com/shiptest-ss13/Shiptest/assets/135062404/7656da59-59ec-40eb-9489-6db58a5a11b8)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
old one stinks, is shit, terrible (and a little dumb)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl: Stiel0248, Positive entropy
imageadd: Resprites the flyswatter
/:cl:


<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->