FOXALIVE - Disables Nanites (#2039)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes nanite research from the tech web pending a more extensive
rework. Does not touch any pre-placed nanite equipment or disks in ruins
and on ships.

## Why It's Good For The Game

nanites are kind of a nightmare for balance, frankly. While they offer
some great potential, a lot of the more Gamerly programs need to be
reworked or axed and that will take time to do.

## Changelog

:cl:
del: Disabled research nodes for nanites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->