Hunter Montagne fix (#2105)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I originally thought that the Hunter Montagne locker not unlocking was
an issue with job overrides in the json. I was super wrong oopsies.
This changes the montagne job and outfit to be a subtype of the hos, and
strips away the hossy stuff that shouldn't be there.

![image](https://github.com/shiptest-ss13/Shiptest/assets/55075747/ec5dae34-80a2-4adb-ad95-deb6b957b895)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The hunter montagne now has proper access.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:phoaly
tweak: tweaks the Hunter Montagne job outfit and job datum
fix: Fixes #2101 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->